+++ expenses/src/codatsyncexpenses/sdkconfiguration.py
@@ -17,7 +17,7 @@ class SDKConfiguration:
     server_idx: int = 0
     language: str = 'python'
     openapi_doc_version: str = 'prealpha'
-    sdk_version: str = '0.31.0'
+    sdk_version: str = '0.31.1'
     gen_version: str = '2.55.0'
 
     def get_server_details(self) -> tuple[str, dict[str, str]]: