+++ expenses/RELEASES.md
@@ -575,3 +575,11 @@ Based on:
 - Speakeasy CLI 1.52.0 (2.55.0) https://github.com/speakeasy-api/speakeasy
 ### Releases
 - [PyPI v0.31.0] https://pypi.org/project/codat-sync-for-expenses/0.31.0 - expenses
+
+## 2023-07-04 00:16:58
+### Changes
+Based on:
+- OpenAPI Doc prealpha https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Expenses.yaml
+- Speakeasy CLI 1.52.0 (2.55.0) https://github.com/speakeasy-api/speakeasy
+### Releases
+- [PyPI v0.31.1] https://pypi.org/project/codat-sync-for-expenses/0.31.1 - expenses